requirements.txt
Django==3.2
django-allauth==0.49.0
django-rest-auth==0.9.5
djangorestframework==3.12.0